requirements.txt
datasets==3.6.0
hydra-core==1.3.2
icecream==2.1.3
loguru==0.7.2
omegaconf==2.3.0
openai==1.85.0
pandas==2.3.0
Pillow==11.2.1
python-dotenv==1.1.0
qwen_vl_utils==0.0.11
Requests==2.32.4
scikit_learn==1.7.0
torch==2.7.0
tqdm==4.67.1
transformers==4.52.4
trl==0.15.2
unsloth==2025.5.9
wandb==0.20.1
